+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 ## [Unreleased]
 ### Added
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
+- Vertical guide at column `N` for `.input` and `.output` files declaring a `// MAX_WIDTH N` comment